ClientRequest Executor fix and log fixes
1) Protocol negotiation request does not have a timeout associated with
it . So if the connection fails the Input/Output buffers are leaked and
cause Out Of Memory errors. Setting the timeout for the Protocol
negotiation requests as the caller expects a timeout.

No new logs are introduced. Only the existing log is made more
informative.

2) When protocol negotiation timesout, an obscure error message is thrown,
IllegalStateException. This happens as the code tries to read the result
in case of timeout. Handle this case and report a meaningful error.

3) Serial request times out on a connection, error message does not
mention the socket that timed out.

4) Fix other minor logging information.